safe writes in multi-user context; implement locks, racefree writes; #12

+++ yacman/yacman.py
@@ -3,12 +3,18 @@ import os
 from collections import Iterable
 import oyaml as yaml
 import logging
+import errno
+import time
+import sys
 from ubiquerg import mkabs
+
 
 _LOGGER = logging.getLogger(__name__)
 
 
 FILEPATH_KEY = "_file_path"
+RO_KEY = "_ro"
+LOCK_PREFIX = "lock."
 
 
 ### Hack for string indexes of both ordered and unordered yaml representations
@@ -62,19 +68,36 @@ class YacAttMap(attmap.PathExAttMap):
     :param str filepath: YAML filepath to the config file.
     """
 
-    def __init__(self, entries=None, filepath=None, yamldata=None):
+    def __init__(self, entries=None, filepath=None, yamldata=None, ro=False, wait_max=100):
 
         if isinstance(entries, str):
             # If user provides a string, it's probably a filename we should read
             # This should be removed at a major version release now that the
             # filepath argument exists, but we retain it for backwards
             # compatibility
-            _LOGGER.debug("The entries argument should be a dict. If you want"
-            " to read a file, use the filepath arg")
+            _LOGGER.debug("The entries argument should be a dict. If you want to read a file, use the filepath arg")
             filepath = entries
             entries = None
 
         if filepath:
+            setattr(self, RO_KEY, ro)
+            # If user provides a string, it's probably a filename we should read
+            # Check if user intends to update the file
+            if not getattr(self, RO_KEY):
+                # attempt to lock the file
+                lock_path = _make_lock_path(filepath)
+                if os.path.exists(lock_path):
+                    raise OSError("The file is locked: {}".format(filepath))
+                else:
+                    try:
+                        _create_file_racefree(lock_path)
+                    except OSError as e:
+                        if e.errno == errno.EEXIST:
+                            # Rare case: file already exists;
+                            # the lock has been created in the split second since the last lock existence check,
+                            # wait for the lock file to be gone, but no longer than `wait_max`.
+                            print("Could not create a lock file, it already exists: {}".format(lock_path))
+                            _wait_for_lock(lock_path, wait_max)
             file_contents = load_yaml(filepath)
             if entries:
                 file_contents.update(entries)
@@ -89,11 +112,29 @@ class YacAttMap(attmap.PathExAttMap):
             setattr(self, FILEPATH_KEY, mkabs(filepath))
 
     def write(self, filename=None):
+        """
+        Writes the contents to a file.
+
+        Makes sure that the object has been created with write capabilities
+
+        :param str filename: a file path to write to
+        :raise OSError: when the object has been created in a read only mode or other process has locked the file
+        :return str: the path to the created file
+        """
+        if hasattr(self, RO_KEY) and getattr(self, RO_KEY):
+            raise OSError("You can't write to a file that was read in RO mode.")
         filename = filename or getattr(self, FILEPATH_KEY)
         if not filename:
             raise Exception("No filename provided.")
+        lock = _make_lock_path(filename)
+        if not hasattr(self, RO_KEY) and os.path.exists(lock):
+            # if the object hasn't been created by reading from file, but a lock exists
+            raise OSError("You can't write to a file that was locked by a different process.")
         with open(filename, 'w') as f:
             f.write(self.to_yaml())
+        if os.path.exists(lock):
+            # remove the lock after writing
+            os.remove(lock)
         return os.path.abspath(filename)
 
     @property
@@ -101,22 +142,78 @@ class YacAttMap(attmap.PathExAttMap):
         """ Most specific type to which an inserted value may be converted """
         return YacAttMap
 
-    def _excl_from_repr(self, k, cls):
-        return k == FILEPATH_KEY
-
     def __repr__(self):
         # Here we want to render the data in a nice way; and we want to indicate
         # the class if it's NOT a YacAttMap. If it is a YacAttMap we just want
         # to give you the data without the class name.
-        return self._render(self._simplify_keyvalue(
-            self._data_for_repr(), self._new_empty_basic_map),
-            exclude_class_list="YacAttMap")
+        return self._render(self._simplify_keyvalue(self._data_for_repr(), self._new_empty_basic_map),
+                            exclude_class_list="YacAttMap")
+
+    def _excl_from_repr(self, k, cls):
+        return k in (FILEPATH_KEY, RO_KEY)
+
+
+def _wait_for_lock(lock_file, wait_max):
+    """
+    Just sleep until the lock_file does not exist or a lock_file-related dynamic recovery flag is spotted
+
+    :param str lock_file: Lock file to wait upon.
+    """
+    sleeptime = .5
+    first_message_flag = False
+    dot_count = 0
+    totaltime = 0
+    while os.path.isfile(lock_file):
+        if first_message_flag is False:
+            print("Waiting for file lock: " + lock_file)
+            first_message_flag = True
+        else:
+            sys.stdout.write(".")
+            dot_count = dot_count + 1
+            if dot_count % 60 == 0:
+                sys.stdout.write("")
+        time.sleep(sleeptime)
+        totaltime += sleeptime
+        sleeptime = min(sleeptime + 2.5, 20)
+        if totaltime >= wait_max:
+            raise RuntimeError("The maximum wait time has been reached")
+
+    if first_message_flag:
+        print("File unlocked")
+
+
+def _create_file_racefree(file):
+    """
+    Creates a file, but fails if the file already exists.
+
+    This function will thus only succeed if this process actually creates the file;
+    if the file already exists, it will cause an OSError, solving race conditions.
+
+    :param str file: File to create.
+    """
+    write_lock_flags = os.O_CREAT | os.O_EXCL | os.O_WRONLY
+    fd = os.open(file, write_lock_flags)
+    os.close(fd)
+    return file
+
+
+def _make_lock_path(lock_name_base):
+    """
+    Create path to lock file with given name as base.
+
+    :param str lock_name_base: Lock file name
+    :return str: Path to the lock file.
+    """
+    base, name = os.path.split(lock_name_base)
+    lock_name = name if name.startswith(LOCK_PREFIX) else LOCK_PREFIX + name
+    return lock_name if not base else os.path.join(base, lock_name)
 
 
 def load_yaml(filename):
     with open(filename, 'r') as f:
         data = yaml.safe_load(f)
     return data
+
 
 def get_first_env_var(ev):
     """
@@ -192,20 +289,3 @@ def select_config(config_filepath=None,
         _LOGGER.error("No configuration file found.")
 
     return selected_filepath
-
-
-def single_folder_writeable(d):
-    return os.access(d, os.W_OK) and os.access(d, os.X_OK)
-
-
-def writeable(outdir, strict_exists=False):
-    """
-    Recursively checks to make sure a folder exists and can be  written to.  
-    """
-    outdir = outdir or "."
-    if os.path.exists(outdir):
-        return _single_folder_writeable(outdir)
-    elif strict_exists:
-        raise MissingFolderError(outdir)
-    return writeable(os.path.dirname(outdir), strict_exists)
-
